Compute a tag one time per invocation backup
If the backup is run and there is no "meta data" file, the one is
generated that contains the "tag" for this invocation of backup. If
the backup is successfully completed, the file is removed. Otherwise,
the next time the backup is invoked, the last-saved tag will be loaded
again. This tag value is used for the ZFS snapshot name and the vault
name on Glacier.

Fixes #11