Use pipe status as helper script exit code
When invoking tar|split via the generated shell script, have the script
use the ${PIPESTATUS[0]} as its exit code, rather than whatever happens
to be the exit code after running tar piped to split. That could very
well be the same thing, but better to be explicit.

rebar3 ct passes