Extract app env values correctly; find bash
When using $PIPESTATUS, we must be sure to use bash, not just any "sh"
that happens to be available as "/bin/sh". As such, look for "bash" in a
fixed set of paths, since bash might not be in the user's PATH,
especially if the process is running as the root user.

rebar3 ct passes