Add a live upload test for the full cycle
Previously the upload tests were faked, such that nothing was actually
uploaded anywhere. While that test is still present, an additional test
that really does upload files has been added. It is triggered by setting
environment variables that inform the test how to operate.

* AKASHITA_LIVE_TEST must be defined (e.g. "1")
* GCS_REGION is set to the Google Cloud location (e.g. "us")
* GCP_PROJECT is set to the Google Cloud project ID
* GCP_CREDENTIALS is set to the path of the credentials file

Also changed all of the tests that use `zfs` to fail abruptly if the
binary cannot be found in the path. Otherwise they appear to succeed
when really nothing is being tested.

Fixes #17

rebar3 ct passes