Generate unique, conformant bucket names
Google Cloud Storage has several restrictions on bucket names, not the
least of which is that they be globally unique. Use a version 1 UUID as
the prefix, and make sure the configured bucket name does not have any
dots. Length is not yet enforced in the application, however.

rebar3 ct passes, including live tests